卡片碎片集齐活动

+++ Core/GObject/Clan.cpp
@@ -4045,7 +4045,7 @@ void   Clan::sendClanCopyInfo(Player * player, UInt8 val)
 
     st << static_cast<UInt16>(_techs->getLev(CLAN_TECH_STATUE));     // 神像科技等级
 //    st << static_cast<UInt16>(_techs->getLev(CLAN_TECH_COPY_LEVEL)); // 副本科技等级
-    st << static_cast<UInt16>(getXianyun()); //仙蕴精华 
+    st << static_cast<UInt32>(getXianyun()); //仙蕴精华
     //st << static_cast<UInt16>(_techs->getLev(CLAN_TECH_COPY_ROB));    // 掠夺科技等级
     st << static_cast<UInt16>(_techs->getLev(CLAN_TECH_COPY_LEVEL)); // 副本科技等级
